cdisc-north-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cdisc-north-engine"
version = "0.1.0"
description = "C-DISC North India GTM outreach automation engine"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "C-DISC Technologies"}]
license = {text = "Proprietary"}
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pandas>=2.1.0",
    "openai>=1.30.0",
    "streamlit>=1.32.0",
    "jinja2>=3.1.0",
    "tenacity>=8.2.0",
]

[project.scripts]
cdisc = "cdisc.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
